// frontend/src/app/home/components/types/user.type.ts
export interface User {
    firstName: string;
    lastName: string;
    address: string;
    city: string;
    state: string;
    pin: string;
    email: string;
    password: string;
}


export interface userLogin {
    email: string;
    password: string;
}

export interface loggedInUser {
    firstName: string;
    lastName: string;
    address: string;
    city: string;
    state: string;
    pin: string;
    email:string;
    udid:string;
}

export interface loginToken {
    token: string;
    expiresInSeconds: number;
    user:loggedInUser;
}